+++ src/apis/organizations.ts
@@ -7,16 +7,21 @@ import { RokkaResponse } from '../response'
 import { State } from '../index'
 
 export interface Organizations {
+  OPTION_PROTECT_DYNAMIC_STACK: string
   get(name: string): Promise<RokkaResponse>
+
   create(
     name: string,
     billingEmail: string,
     displayName: string
   ): Promise<RokkaResponse>
+
+  setOption(organizationName:string, name:string, value:boolean|string): Promise<RokkaResponse>
 }
 
 export default (state: State) => {
   const organizations: Organizations = {
+    OPTION_PROTECT_DYNAMIC_STACK: 'protect_dynamic_stack',
     /**
      * Get a list of organizations.
      *
@@ -54,7 +59,12 @@ export default (state: State) => {
         billing_email: billingEmail,
         display_name: displayName
       })
+    },
+
+    setOption: (organizationName, name, value) => {
+      return state.request('PUT', `organizations/${organizationName}/options/${name}`, value)
     }
+
   }
 
   return {